requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
mysql-connector-python==8.1.0
numpy==1.25.2
pandas==2.1.0
protobuf==4.21.12
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
shiboken6==6.5.2
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
